Finished

+++ bid.h
@@ -6,8 +6,8 @@
 struct Bid {
   int id;
   std::vector<int> regions;
-  float payment;
-  Bid(std::vector<int> regions, float payment, int id)
+  double payment;
+  Bid(std::vector<int> regions, double payment, int id)
       : regions(regions), payment(payment), id(id) {}
 };
